Commit 8
Linking the list to the logic analysis process and improving overall user interaction process with a more clear and easy to understand terminal output.

The problem right now is that I cant figure out why it only does analysis for the last padawan that was inputted into the list, and not all of the previous. Will try to fix it in the next commit.