Commit #18
Added a new class for linear search to be able to look for Logically ready and force sentisitve padawans.

added a evaluate_all function to clean up the code. This function will allow to call logical evaluation with one simple line

Adapted linear search to be able to use new class

Added validation to the input system.

Fixed some lines for overall code and output clearence